+++ src/components/ids-notification-banner/ids-notification-banner.ts
@@ -81,7 +81,7 @@ export default class IdsNotificationBanner extends Base {
         <div class="ids-notification-banner-button" part="button">
           <ids-button type="tertiary">
             <span class="audible">Close Button</span>
-            <ids-icon slot="icon" icon="close" size="small"></ids-icon>
+            <ids-icon icon="close" size="small"></ids-icon>
           </ids-button>
         </div>
       </div>