+++ src/components/ids-pager/ids-pager-dropdown.ts
@@ -53,7 +53,7 @@ export default class IdsPagerDropdown extends Base {
     return `
       <div class="ids-pager-dropdown">
         <ids-menu-button id="pager-size-menu-button" menu="pager-size-menu" role="button" dropdown-icon>
-          <span slot="text">${this.pageSize} ${this.label}</span>
+          <span>${this.pageSize} ${this.label}</span>
         </ids-menu-button>
         <ids-popup-menu id="pager-size-menu" target="#pager-size-menu-button" trigger-type="click">
           ${this.#itemsTemplate()}